Add proper printf, printf memory map

+++ kernel/arch/x86/paging.c
@@ -4,12 +4,13 @@
 
 #define TLB_FLUSH(addr) __asm__ volatile ("invlpg (%0)" :: "r" (virt) : "memory")
 
-int8 map_page_large (void * virt, struct pd_entry_s pd_entry)
+int8_t map_page_large (void * virt, struct pd_entry_s pd_entry)
 {
-	if (page_dir[(unsigned long)virt >> 22].present)
+	if (page_dir[(uint32_t)virt >> 22].present)
 		return -1;
+	/**/
 
-	page_dir[(unsigned long)virt >> 22] = pd_entry;
+	page_dir[(uint32_t)virt >> 22] = pd_entry;
 	TLB_FLUSH(virt);
 	return 0;
 }
